feat: add BuiltInContainer widget and update exports
